backend/requirements.txt
fastapi
uvicorn[standard]
reportlab
docxtpl
